+++ node/resolvers/catalog/index.ts
@@ -514,6 +514,7 @@ export const queries = {
       translatedArgs,
       searchMetaData,
       productsRaw,
+      pageInfo: { from: args.from, to: args.to }
     }
   },